+++ api/src/controllers/post.controller.ts
@@ -31,7 +31,7 @@ const postController: PostController = {
 
     // User is authenticated before reaching this endpoint
     const user = req.user as User; // Type assertion
-    const { id } = user;
+    const { id, username } = user;
     /* const newPost = await createPost({
       blogId: user.blog.id,
       authorId: user.id,