+++ api/src/interfaces/request.d.ts
@@ -13,5 +13,6 @@ declare module "express-serve-static-core" {
   interface Request {
     user?: User;
     accessToken?: string;
+    refreshToken?: string;
   }
 }